requirements.txt
alembic==1.7.7
aniso8601==6.0.0
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.4
distlib==0.3.4
filelock==3.7.0
Flask==1.0.3
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.2.1
MarkupSafe==1.1.1
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.8.2
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2019.1
requests==2.28.1
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.26.11
Werkzeug==0.15.5
